+++ Runtime/Extensions/StringExtensions.cs
@@ -1,0 +1,10 @@
+namespace OcUtility
+{
+    public static class StringExtensions
+    {
+        public static bool equal(this string source, string value)
+        {
+            return string.CompareOrdinal(source, value) == 0;
+        }
+    }
+}